+++ src/jml.js
@@ -1049,6 +1049,13 @@ const jml = function jml (...args) {
         // Handled during element creation
         break;
       } case '$custom': {
+        if (
+          attVal !== null &&
+          (typeof attVal === 'object' || typeof attVal === 'function') &&
+          Object.prototype.propertyIsEnumerable.call(attVal, '__proto__')
+        ) {
+          throw new TypeError('`$custom` may not define `__proto__`');
+        }
         Object.assign(elem, attVal);
         break;
       } case '$define': {